fixes unit test
